move the saramconfig from kafkaconfig to config.go
