+++ topics.go
@@ -9,7 +9,6 @@ import (
 	"math/rand"
 
 	"github.com/Shopify/sarama"
-	"github.com/damiannolan/sasl/oauthbearer"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -70,90 +69,6 @@ func createTlsConfig(CAPath string, SkipVerify bool) *tls.Config {
 
 }
 
-func SaramaConfigFromKafkaConfig(conf KafkaConfig) *sarama.Config {
-	config := sarama.NewConfig()
-	config.Metadata.Full = true
-	config.Net.TLS.Enable = conf.SSL.Enabled
-	if conf.Krb5.Enabled {
-		config.Net.SASL.Enable = true
-		config.Net.SASL.Mechanism = sarama.SASLTypeGSSAPI
-		config.Net.SASL.GSSAPI.Realm = conf.Krb5.Realm
-		config.Net.SASL.GSSAPI.Username = conf.Krb5.Username
-		if conf.Krb5.Keytab != "" {
-			config.Net.SASL.GSSAPI.AuthType = sarama.KRB5_KEYTAB_AUTH
-			config.Net.SASL.GSSAPI.KeyTabPath = conf.Krb5.Keytab
-
-		} else {
-			config.Net.SASL.GSSAPI.AuthType = sarama.KRB5_USER_AUTH
-			config.Net.SASL.GSSAPI.Password = conf.Krb5.Password
-		}
-		if conf.Krb5.ServiceName == "" {
-			config.Net.SASL.GSSAPI.ServiceName = "kafka"
-		} else {
-			config.Net.SASL.GSSAPI.ServiceName = conf.Krb5.ServiceName
-		}
-		if conf.Krb5.KerberosConfigPath == "" {
-			config.Net.SASL.GSSAPI.KerberosConfigPath = "/etc/krb5.conf"
-		} else {
-			config.Net.SASL.GSSAPI.KerberosConfigPath = conf.Krb5.KerberosConfigPath
-		}
-	}
-	// Set Sasl plain if configured
-	if conf.SaslPlain.Enabled {
-
-		config.Net.SASL.Enable = true
-		config.Net.SASL.Mechanism = sarama.SASLTypePlaintext
-		config.Net.SASL.User = conf.SaslPlain.Username
-		config.Net.SASL.Password = conf.SaslPlain.Password
-	}
-	// Token auth if configured
-	if conf.TokenAuth.Enabled {
-		if conf.TokenAuth.ClientID == "" {
-			log.Fatal("required 'client' field  not defined in tokenauth section")
-		}
-		if conf.TokenAuth.Secret == "" {
-			log.Fatal("required 'secret' field  not defined in tokenauth section")
-		}
-		if conf.TokenAuth.TokenUrl == "" {
-			log.Fatal("required 'url' field  not defined in tokenauth section")
-		}
-		config.Net.SASL.Enable = true
-		config.Net.SASL.Mechanism = sarama.SASLTypeOAuth
-		// We use a special handler if Confluent Metadata service tokens are used.
-		if conf.TokenAuth.IsConfluentMDS {
-			config.Net.SASL.TokenProvider = NewTokenProviderConfluentMDS(conf.TokenAuth.ClientID, conf.TokenAuth.Secret, conf.TokenAuth.TokenUrl)
-
-		} else {
-			config.Net.SASL.TokenProvider = oauthbearer.NewTokenProvider(conf.TokenAuth.ClientID, conf.TokenAuth.Secret, conf.TokenAuth.TokenUrl)
-		}
-	}
-	if conf.SSL.Enabled && (conf.SSL.CA != "" || conf.SSL.SkipVerify) {
-		tlsConfig := createTlsConfig(conf.SSL.CA, conf.SSL.SkipVerify)
-		config.Net.TLS.Config = tlsConfig
-	}
-	if conf.Producer.MaxMessageBytes != 0 {
-		config.Producer.MaxMessageBytes = conf.Producer.MaxMessageBytes
-	}
-	if conf.Producer.Compression != "" {
-		switch conf.Producer.Compression {
-		case "snappy":
-			config.Producer.Compression = sarama.CompressionSnappy
-		case "gzip":
-			config.Producer.Compression = sarama.CompressionGZIP
-		case "lz4":
-			config.Producer.Compression = sarama.CompressionLZ4
-		case "zstd":
-			config.Producer.Compression = sarama.CompressionZSTD
-		case "none":
-			config.Producer.Compression = sarama.CompressionNone
-		}
-	} else {
-		// Use snappy as default if none other is specified
-		config.Producer.Compression = sarama.CompressionSnappy
-	}
-	return config
-
-}
 func NewKafkaAdmin(conf KafkaConfig) KafkaAdmin {
 
 	var admin KafkaAdmin
